fixed delete file

+++ src/StorageFiles.php
@@ -21,7 +21,6 @@ final class StorageFiles
             {
                 $this->deleteFromCache($file->path);
                 $file->delete();
-                StorageFileContents::where('file_id', $file->file_id)->delete();
             }
         }
     }
@@ -75,18 +74,12 @@ final class StorageFiles
         {
             $this->deleteFromCache($pathOrModel->path);
             $pathOrModel->delete();
-            StorageFileContents::where('file_id', $pathOrModel->file_id)->delete();
         }
         else
         {
             $this->deleteFromCache($pathOrModel);
             $query = StorageFile::where('path', $pathOrModel);
-            if($query->exists()) 
-            {
-                $fileId = $query->value('file_id');
-                $query->delete();
-                StorageFileContents::where('file_id', $fileId)->delete();
-            }
+            if($query->exists()) $query->delete();
         }
     }
